Combined dashboard tabs into 1 page to stop screen flashing on tab switches
