+++ components/dashboard/Dashboard.module.css
@@ -1,0 +1,3 @@
+.tab {
+  margin-left: 12px;
+}
