+++ pages/signin/index.tsx
@@ -41,7 +41,7 @@ export default function Signin({ session }: SignUpProps) {
     const user: any = newSession.user;
 
     // Need to refresh CSP
-    window.location.assign(user.isBusiness ? "/dashboard" : "/");
+    window.location.assign(user.isBusiness ? "/dashboard?tab=inventory" : "/");
   };
 
   const onProviderSignIn = (provider: string) => {
@@ -52,7 +52,7 @@ export default function Signin({ session }: SignUpProps) {
     const user: any = session.user;
 
     // Need to refresh CSP
-    window.location.assign(user.isBusiness ? "/dashboard" : "/");
+    window.location.assign(user.isBusiness ? "/dashboard?tab=inventory" : "/");
     return null;
   }
